+++ src/app/components/embedded-content-space-transfers.js
@@ -84,7 +84,7 @@ export default OneproviderEmbeddedContainer.extend(...mixins, {
     changeListTab(tab) {
       return this.get('navigationState').changeRouteAspectOptions({
         tab,
-      });
+      }, true);
     },
   },
 });